requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
Flask==2.3.3
Jinja2==3.1.2
python-dotenv==1.0.0
lxml_html_clean 